feat: add langfuse tracing to manipulation bench moved tracing code to result_processing/
